+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+## Added
+- Support for multiple p5 sketches.
 
 ## [0.3.0] - 2017-04-21
 ### Added